Turn displayPasteError helper into hook
