+++ assets/src/stories-editor/blocks/amp-story-page/copy-paste-handler.js
@@ -13,7 +13,7 @@ import { withDispatch, useSelect, useDispatch } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import { copyTextToClipBoard, isPageBlock, useIsBlockAllowedOnPage, displayPasteError } from '../../helpers';
+import { copyTextToClipBoard, isPageBlock, useIsBlockAllowedOnPage, useDisplayPasteError } from '../../helpers';
 
 function CopyPasteHandler( { children, onCopy, clientId, isSelected } ) {
 	const {
@@ -35,8 +35,8 @@ function CopyPasteHandler( { children, onCopy, clientId, isSelected } ) {
 	);
 
 	const { insertBlock } = useDispatch( 'core/block-editor' );
-	const { createErrorNotice } = useDispatch( 'core/notices' );
 	const isBlockAllowedOnPage = useIsBlockAllowedOnPage();
+	const displayPasteError = useDisplayPasteError();
 
 	const onPaste = ( event ) => {
 		// Ignore if the Page is not the selected page.
@@ -83,7 +83,7 @@ function CopyPasteHandler( { children, onCopy, clientId, isSelected } ) {
 			if ( isBlockAllowedOnPage( pastedBlock.name, clientId ) ) {
 				insertBlock( pastedBlock, blocksOnPage.length, clientId );
 			} else {
-				displayPasteError( pastedBlock.name, createErrorNotice );
+				displayPasteError( pastedBlock.name );
 			}
 		} );
 	};
