+++ assets/src/stories-editor/helpers/index.js
@@ -60,4 +60,4 @@ export { default as getRelativeElementPosition } from './getRelativeElementPosit
 export { default as isCTABlock } from './isCTABlock';
 export { default as useMoveBlockToPage } from './useMoveBlockToPage';
 export { default as useIsBlockAllowedOnPage } from './useIsBlockAllowedOnPage';
-export { default as displayPasteError } from './displayPasteError';
+export { default as useDisplayPasteError } from './useDisplayPasteError';
